requirements.txt
numpy
librosa
soundfile
Pillow
sounddevice
pygame